vscode-ext: split into multiple files
